requirements.txt
numpy<=1.23.5
torch==1.13.1
# dataclasses==0.6
tensorboard==2.11.0
tqdm==4.64.1
matplotlib==3.6.2
hydra-core==1.3.1
omegaconf==2.3.0
wandb==0.13.9
dotmap==1.3.30
opencv-python==4.7.0.68